@@ -21,7 +21,7 @@ setup(
     long_description=long_desc,
     long_description_content_type='text/markdown',
     install_requires=install_requires,
-    python_requires=">=3.8.*",
+    python_requires=">=3.10",
     classifiers=[
         "Development Status :: 3 - Alpha",
         "Intended Audience :: Science/Research",
