@@ -10,7 +10,7 @@ from ase.calculators.psi4 import Psi4
 from fff.simulation.utils import write_to_string, read_from_string
 
 
-def run_calculator(xyz: str, calc: Calculator | dict, temp_path: Optional[str] = None) -> str:
+def run_calculator(xyz: str, calc: Calculator | dict, temp_path: Optional[str] = None, subprocess: bool = True, timeout: Optional[float] = None) -> str:
     """Run an NWChem computation on the requested cluster
 
     Args:
@@ -18,14 +18,19 @@ def run_calculator(xyz: str, calc: Calculator | dict, temp_path: Optional[str] =
         calc: ASE calculator to use. Either the calculator object or a dictionary describing the settings
             (only Psi4 supported at the moment for dict)
         temp_path: Base path for the scratch files
+        subprocess: Whether to run computation in a subprocess. Useful for calculators which do not exit cleanly
+        timeout: How long to wait for result, used only if subprocess is True
     Returns:
         Atoms after the calculation in a JSON format
     """
 
     # Some calculators do not clean up their resources well
-    with ProcessPoolExecutor(max_workers=1) as exe:
-        fut = exe.submit(_run_calculator, str(xyz), calc, temp_path)  # str ensures proxies are resolved
-        return fut.result()
+    if subprocess:
+        with ProcessPoolExecutor(max_workers=1) as exe:
+            fut = exe.submit(_run_calculator, str(xyz), calc, temp_path)  # str ensures proxies are resolved
+            return fut.result(timeout=timeout)
+    else:
+        return _run_calculator(str(xyz), calc, temp_path)
 
 
 def _run_calculator(xyz: str, calc: Calculator | dict, temp_path: Optional[str] = None) -> str:
@@ -34,23 +39,27 @@ def _run_calculator(xyz: str, calc: Calculator | dict, temp_path: Optional[str] 
     # Parse the atoms object
     atoms = read_from_string(xyz, 'xyz')
 
+    start_dir = os.getcwd()
     with TemporaryDirectory(dir=temp_path, prefix='fff') as temp_dir:
-        # Execute from the temp so that the calculators do not interfere
-        os.chdir(temp_dir)
-
-        # Special case for Psi4 which sets the run directory on creating the object
-        if isinstance(calc, dict):
-            calc = calc.copy()
-            assert calc.pop('calc') == 'psi4', 'only psi4 is supported for now'
-            calc = Psi4(**calc, directory=temp_dir, PSI_SCRATCH=temp_path)
-
-        # Run the calculation
-        atoms.calc = calc
         try:
-            atoms.get_forces()
-            atoms.get_potential_energy()
-        except BaseException as exc:
-            raise ValueError(f'Calculation failed: {exc}')
+            # Execute from the temp so that the calculators do not interfere
+            os.chdir(temp_dir)
 
-        # Convert it to JSON
-        return write_to_string(atoms, 'json')
+            # Special case for Psi4 which sets the run directory on creating the object
+            if isinstance(calc, dict):
+                calc = calc.copy()
+                assert calc.pop('calc') == 'psi4', 'only psi4 is supported for now'
+                calc = Psi4(**calc, directory=temp_dir, PSI_SCRATCH=temp_path)
+
+            # Run the calculation
+            atoms.calc = calc
+            try:
+                atoms.get_forces()
+                atoms.get_potential_energy()
+            except BaseException as exc:
+                raise ValueError(f'Calculation failed: {exc}')
+
+            # Convert it to JSON
+            return write_to_string(atoms, 'json')
+        finally:
+            os.chdir(start_dir)
